syntactic adjustment

+++ src/main/java/kz/halykacademy/bookstore/service/impl/AuthorServiceImpl.java
@@ -2,12 +2,10 @@ package kz.halykacademy.bookstore.service.impl;
 
 import kz.halykacademy.bookstore.dto.*;
 import kz.halykacademy.bookstore.entity.Author;
-import kz.halykacademy.bookstore.entity.Books;
 import kz.halykacademy.bookstore.errors.ResourceNotFoundeException;
 import kz.halykacademy.bookstore.mapper.AuthorMapper;
 import kz.halykacademy.bookstore.repository.AuthorRepository;
 import kz.halykacademy.bookstore.repository.BookRepository;
-import kz.halykacademy.bookstore.repository.GenreRepository;
 import kz.halykacademy.bookstore.service.AuthorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,11 +17,8 @@ import java.util.function.Supplier;
 @RequiredArgsConstructor
 public class AuthorServiceImpl implements AuthorService {
 
-
     private final AuthorRepository authorRepository;
-    private final BookRepository bookRepository;
-    private  final GenreRepository genreRepository;
-    private  final AuthorMapper authorMapper;
+    private final AuthorMapper authorMapper;
 
 
     @Override
@@ -33,16 +28,15 @@ public class AuthorServiceImpl implements AuthorService {
                 .stream()
                 .map(authorMapper::toDTO)
                 .toList();
-
     }
 
     @Override
-    public AuthorDTO getAuthorById(long authorId) throws Throwable {
+    public AuthorDTO getAuthorById(Long authorId) throws Throwable {
 
         return authorRepository.findById(authorId)
                 .map(authorMapper::toDTO)
                 .orElseThrow((Supplier<Throwable>) () ->
-                        new ResourceNotFoundeException("Author %s not found".formatted(authorId)));
+                        new ResourceNotFoundeException("Author with id %s not found".formatted(authorId)));
     }
 
 
@@ -50,34 +44,39 @@ public class AuthorServiceImpl implements AuthorService {
     public AuthorDTO addAuthor(SaveAuthorDTO author) {
         Author saved = authorRepository.save(
                 new Author(
-                        author.getAuthorId(),
-                        author.getFull_name(),
-                        author.getDate_of_birth(),
-                        null
+                        null,
+                        author.getFullName(),
+                        author.getDateOfBirth(),
+                        null,
+                        false
                 )
         );
-        return  authorMapper.toDTO(saved);
+        return authorMapper.toDTO(saved);
     }
 
     @Override
-    public AuthorDTO updateAuthor(SaveAuthorDTO authorDTO, long id) {
-        Author author = authorRepository.findById(id).get();
+    public AuthorDTO updateAuthor(UpdateAuthorDTO authorDTO) throws Throwable {
+        Author author = authorRepository.findById(authorDTO.getId()).orElseThrow((Supplier<Throwable>) () ->
+                new ResourceNotFoundeException("Author with id %s not found".formatted(authorDTO.getId())));
 
         Author updateAuthor = authorRepository.save(
                 new Author(
-                        author.getAuthorId(),
-                        authorDTO.getFull_name(),
-                        authorDTO.getDate_of_birth(),
-                        author.getBooks()
+                        authorDTO.getId(),
+                        authorDTO.getFullName(),
+                        authorDTO.getDateOfBirth(),
+                        author.getBooks(),
+                        author.isDeleted()
                 )
         );
         return authorMapper.toDTO(updateAuthor);
     }
 
     @Override
-    public void deleteAuthor(long authorId) throws Exception {
-
-       authorRepository.deleteById(authorId);
+    public void deleteAuthor(Long authorId) {
+        if(!authorRepository.existsByAuthorId(authorId)){
+            throw new ResourceNotFoundeException("Author with id %s not found".formatted(authorId));
+        }
+        authorRepository.deleteById(authorId);
     }
 
     @Override
@@ -85,21 +84,11 @@ public class AuthorServiceImpl implements AuthorService {
         return authorRepository.findByName(name).stream().map(authorMapper::toDTO).toList();
     }
 
-    @Override
-    public AuthorDTO addBook(long authorId, long bookId) {
-
-
-        Author author = authorRepository.findById(authorId).get();
-        Books books = bookRepository.findById(bookId).get();
-        author.getBooks().add(books);
-
-        return authorMapper.toDTO(authorRepository.save(author));
-    }
 
     @Override
     public List<AuthorGenreListDTO> findAuthorsByGenreList(List<String> genreList) {
-        return  authorRepository.getAuthorByGenreList(genreList).stream().map(authorMapper::toGenreDTO).toList();
-    }
+        return authorRepository.getAuthorByGenreList(genreList).stream().map(authorMapper::toGenreDTO).toList();
 
+    }
 
 }
